Make public class "Node" private nested class of "IntLinkedInt" class
